+++ iModelBridgeCore/Pointools/PointoolsVortexAPI/src/pt/datatreeIO.cpp
@@ -111,7 +111,7 @@ struct WriteNodeVisitor
 	// visitor functors
 	void operator() (const NodeID &nid, const Node *node)
 	{	
-		uint8 type_index = node->typeId();
+		uint8 type_index = (uint8) node->typeId();
 
 		// write identifier	
 		writeId( nid );
